requirements.txt
chromadb==0.5.3
datasets==2.20.0
langchain==0.2.12
langchain_chroma==0.1.3
langchain_community==0.2.11
langchain_core==0.2.28
langchain_huggingface==0.0.3
langchain_openai==0.1.20
langchain_text_splitters==0.2.2
portkey_ai==1.7.0
PyYAML==6.0.1
Requests==2.32.3
streamlit==1.37.0
streamlit_authenticator==0.3.3
stackapi==0.3.1
langchainhub==0.1.21
pypdf==4.3.1
jinja2>=3.1.0
unstructured==0.15.7